added test mode

+++ logtrimmer.py
@@ -6,7 +6,7 @@ import sys
 import re
 
 
-def trim_log_files():
+def trim_log_files(test=False):
 
     config = yaml.load(open('config.yaml', 'r'))
 
@@ -28,11 +28,27 @@ def trim_log_files():
         print "Undefined logging type: ", config['sort']
         sys.exit(1)
 
+    if test:
+        print "Log files found are: "
+        print files
+
     while len(files) > config['history']:
         target = files.pop(0)
-        os.remove(target)
+        if test:
+            print "Would remove file: ", target
+        else:
+            os.remove(target)
 
 
 if __name__ == "__main__":
 
+    print sys.argv
+
+    if len(sys.argv) > 1:
+        if sys.argv[1].lower() == 'test':
+            print "Running in test mode"
+            trim_log_files(test=True)
+            sys.exit(0)
+
     trim_log_files()
+    sys.exit(0)
